+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handler/HandlerMappingException.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.handler;
 
 public class HandlerMappingException extends RuntimeException {
