+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/viewresolver/JsonViewResolver.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.viewresolver;
 
 import webmvc.org.springframework.web.servlet.View;
 import webmvc.org.springframework.web.servlet.view.JsonView;
