+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/viewresolver/ViewResolverException.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.viewresolver;
 
 public class ViewResolverException extends RuntimeException {
