+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/tobe/AnnotationHandlerAdapterTest.java
@@ -11,6 +11,9 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.mvc.adapter.AnnotationHandlerAdapter;
+import webmvc.org.springframework.web.servlet.mvc.handler.HandlerExecution;
+import webmvc.org.springframework.web.servlet.mvc.handler.HandlerMapping;
 import webmvc.org.springframework.web.servlet.view.JspView;
 
 class AnnotationHandlerAdapterTest {
